Implemented the last menu commands (instances and log) \nAdded: Test for APlugin\nModified: APlugin is now responsible for generating it's own instanceID

+++ Hydra.Commons/src/com/eofstudio/hydra/commons/plugin/IPlugin.java
@@ -4,8 +4,8 @@ import java.io.IOException;
 
 public interface IPlugin extends Runnable
 {
-	int getPluginID();
-	int getInstanceID();
+	long getPluginID();
+	long getInstanceID();
 	void addConnection( IHydraPacket packet ) throws IOException;
 	Thread getThread();
 }
